refactor(mp-tui): remove utils module and integrate theme directly in renderer
